+++ signal_analog/flow.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """This module provides bindings for the SignalFx SignalFlow DSL."""
-
+from enum import Enum
 from numbers import Number
 
 from six import string_types
@@ -68,12 +68,12 @@ class Program(object):
         Arguments:
             stmt: the statement to validate
         """
-        if not stmt or not issubclass(stmt.__class__, Function):
+        if not stmt or (not issubclass(stmt.__class__, Function) and not issubclass(stmt.__class__, Plot)):
             msg = "Attempted to build a program with something other than " +\
                    "SignalFlow statements. Received '{0}' but expected a " +\
-                   "{1}"
+                   "{1} or {2}"
             raise ValueError(msg.format(
-                stmt.__class__.__name__, Function.__name__))
+                stmt.__class__.__name__, Function.__name__, Plot.__name__))
 
     def add_statements(self, *statements):
         """Add a statement to this program.
@@ -123,6 +123,10 @@ class Program(object):
         """Validate that at least 1 statement is published for this Program."""
         def find_publish(statement):
 
+            # Inspect the internals of the Plot
+            if isinstance(statement, Plot):
+                statement = statement.plot
+
             # Inspect the left hand side of the assignment
             if isinstance(statement, Assign):
                 statement = statement.expr
@@ -140,6 +144,67 @@ class Program(object):
 
         if len(publish_statements) < 1:
             raise ProgramDoesNotPublishTimeseriesError(statements)
+
+
+class Plot(object):
+    """Represents a Plot for a Chart as configured in the SignalFx UI.  It is a helper class that makes it simpler
+       to use lower-level abstractions like Assign, Function, Data, publish()"""
+
+    def __init__(self, assigned_name, signal_name, filter=None, rollup=None, fx=None, label=None, visible=True,
+                 extrapolation_policy=None, max_extrapolations=None):
+        """Represents a Plot for a Chart as configured in the SignalFx UI.  It is a helper class that makes it simpler
+        to use lower-level abstractions like Assign, Function, Data, publish()
+
+        Other options available in the UI for a plot include: display_units, plot_color, and visualization_type.  These
+        values can be configured in with_publish_label_options() in the Chart itself.  See PublishLabelOptions.
+
+        Example:
+
+        >>> TimeSeriesChart() \
+        >>>    .with_name("Cpu Utilization") \
+        >>>    .with_program(
+        >>>       Plot("A", "CPUUtilization", filter, rollup="max", fx=[Mean(by="app")])
+        >>> )
+
+        Arguments:
+            assigned_name: the assigned name used in the SignalFlow program, use A-Z to keep compatibility with the
+                    builder in the SignalFx UI. E.g. the "A" in "A = data('metric_name')"
+            signal_name: the metric name, e.g. "CPUUtilization"
+            filter: the filter to apply to the metric, e.g. And(Filter("env, "prod"), Filter("app", "foo"))
+            rollup: RollupType If None then the default rollup for the metric is used. Otherwise use the RollupType
+                    enum. E.g. RollupType.average, RollupType.rate, RollupType.delta, RollupType.max
+            fx: List of Signal Flow function(s) to apply e.g. [ Mean(by="app") ]
+            label: the name visible in the chart on hover.  This is also the label used in PublishLabelOptions
+            visible: True if this plot should be visible in the chart.  False is used for values that used in a formula
+                    but that aren't displayed themselves.  Defaults to True.
+            extrapolation_policy: String How to extrapolate missing data. One of the following string values:
+
+                    null: Missing data is not emitted. (default)
+
+                    zero: Missing data is considered to be a value of 0
+
+                    last_value: Missing data is considered to have the last value
+            max_extrapolations: Int How many extrapolations will be performed when data is no longer received from a
+                    source. A negative value indicates infinite extrapolation.
+        """
+        if not assigned_name:
+            raise ValueError("assigned_name is required, e.g. 'A', 'B'")
+
+        if not signal_name:
+            raise ValueError("signal_name is required, e.g. 'CPUUtilization'")
+
+        data = Data(signal_name, filter, rollup, extrapolation_policy, max_extrapolations)
+
+        if fx and not isinstance(fx, list):
+            raise ValueError("fx should be of type list")
+        elif fx:
+            data.call_stack = fx
+
+        assign = Assign(assigned_name, data.publish(label, visible))
+        self.plot = assign
+
+    def __str__(self):
+        return self.plot.__str__()
 
 
 class Function(object):
@@ -379,9 +444,11 @@ class Function(object):
         """Publish the output of a stream so that it is visible outside of a
            computation.
 
+           This is also the label used in the 'PublishLabelOptions' class.
+
         Arguments:
             label: String defining a label for the stream
-            enable: Boolean
+            enable: Boolean True/False means show/hide the Time Series in the chart.
         """
         self.call_stack.append(Publish(label=label, enable=enable))
         return self
@@ -599,6 +666,67 @@ class Function(object):
         return self
 
 
+class RollupType(Enum):
+    """The Roll-up Type for SignalFlow
+
+    See [SignalFx Documentation on Roll-ups](https://docs.signalfx.com/en/latest/charts/resolution-rollups.html)
+    """
+
+    def __str__(self):
+        """
+        Convert to string representation expected by SignalFx
+        """
+        return '"' + self.value + '"'
+
+    average = "average"
+    """
+    Default for gauges. Divide the result of Sum by the count of datapoints for the time interval.
+    """
+
+    count = "count"
+    """
+    The number of datapoints observed in the time interval.
+    """
+
+    delta = "delta"
+    """
+    Return the difference between the first and last value observed in the time interval. This difference is never
+    negative. If the value of a cumulative counter datapoint is ever smaller than the previous value, the delta for
+    that interval will be just the new value, not the negative difference between them.
+    """
+
+    latest = "latest"
+    """
+    Return the value of the last datapoint received in the time interval.
+    """
+
+    max = "max"
+    """
+    Select the maximum value seen in the time interval.
+    """
+
+    min = "min"
+    """
+    The minimum value seen in the time interval.
+    """
+
+    rate = "rate"
+    """
+    Default for counters.  Divide the result of Sum (counter) or Delta (cumulative counter) by the number of seconds
+    in the time interval.
+    """
+
+    sum = "sum"
+    """
+    Compute the sum of all points for the time interval.
+    """
+
+    lag = "lag"
+    """
+    Return the average time in milliseconds between each datapoint’s timestamp and the time of its receipt at SignalFx.
+    """
+
+
 class StreamMethod(object):
 
     def __init__(self, name):
@@ -688,29 +816,14 @@ class Data(Function):
                  rollup=None, extrapolation=None, maxExtrapolations=None):
         """The data() function is used to create a stream.
 
+        Assigning data is recommended to keep your Signal Flow program compatible with the SignalFx UI builder,
+        e.g. Assign('A', Data('mymetric', filter).publish(label)).
+
         Arguments:
             metric: String metric name (can use * as a wildcard)
             filter: String filter name to match
-            rollup: String If None then the default rollup for the metric is used. Otherwise one of the following
-                    string values:
-
-                    average (default for gauges)
-
-                    count (useful to tell how many points were received in a timeslice)
-
-                    delta (useful for cumulative counters)
-
-                    latest (useful for gauges)
-
-                    max (useful for cumulative counters)
-
-                    min (useful for gauges)
-
-                    rate (default for cumulative and distributed counters)
-
-                    sum (useful for distributed counters)
-
-                    lag (the measured ingest lag for the timeseries for each timeslice)
+            rollup: RollupType If None then the default rollup for the metric is used. Otherwise use the RollupType
+                    enum. E.g. RollupType.average, RollupType.rate, RollupType.delta, RollupType.max
             extrapolation: String How to extrapolate missing data. One of the following string values:
 
                             null: Missing data is not emitted. (default)
@@ -772,26 +885,8 @@ class Graphite(Function):
         Arguments:
             metric: String metric name (can use * as a wildcard)
             filter: String filter name to match
-            rollup: String If None then the default rollup for the metric is used. Otherwise one of the following
-                    string values:
-
-                    average (default for gauges)
-
-                    count (useful to tell how many points were received in a timeslice)
-
-                    delta (useful for cumulative counters)
-
-                    latest (useful for gauges)
-
-                    max (useful for cumulative counters)
-
-                    min (useful for gauges)
-
-                    rate (default for cumulative and distributed counters)
-
-                    sum (useful for distributed counters)
-
-                    lag (the measured ingest lag for the timeseries for each timeslice)
+            rollup: RollupType If None then the default rollup for the metric is used. Otherwise use the RollupType
+                    enum. E.g. RollupType.average, RollupType.rate, RollupType.delta, RollupType.max
             extrapolation: String How to extrapolate missing data. One of the following string values:
 
                             null: Missing data is not emitted. (default)
@@ -823,26 +918,8 @@ class Newrelic(Function):
         Arguments:
             metric: String metric name (can use * as a wildcard)
             filter: String filter name to match
-            rollup: String If None then the default rollup for the metric is used. Otherwise one of the following
-                    string values:
-
-                    average (default for gauges)
-
-                    count (useful to tell how many points were received in a timeslice)
-
-                    delta (useful for cumulative counters)
-
-                    latest (useful for gauges)
-
-                    max (useful for cumulative counters)
-
-                    min (useful for gauges)
-
-                    rate (default for cumulative and distributed counters)
-
-                    sum (useful for distributed counters)
-
-                    lag (the measured ingest lag for the timeseries for each timeslice)
+            rollup: RollupType If None then the default rollup for the metric is used. Otherwise use the RollupType
+                    enum. E.g. RollupType.average, RollupType.rate, RollupType.delta, RollupType.max
             extrapolation: String How to extrapolate missing data. One of the following string values:
 
                             null: Missing data is not emitted. (default)
@@ -956,8 +1033,10 @@ class Assign(Function):
     def __init__(self, assignee, expr):
         """Assign the given expression to the assignee
 
+           A-Z are recommended values to keep your Signal Flow program compatible with
+           the SignalFx UI builder.
         Arguments:
-            assignee: the name to which to assign the expression
+            assignee: the name to which to assign the expression, e.g. 'A', 'B'
             expr: the expression to assign
 
         Returns: